requirements.txt
sqlalchemy[asyncio]
asyncpg
sqlalchemy-utils
fastapi[all]
uvicorn[standart]
alembic
python-jose[cryptogaphy]
